Module 4
* Initial structure

* Added CRUD operations for GiftCertificate and Tag

* Added search and sort operations

* Added enum parameters for controllers

* Added exception handling & docs

* Added unit test for services

* Added docs for controllers

* Added dto & refactoring

* Added more tests & refactoring

* Added integration tests with hsqldb & refactoring

* Refactoring

* Refactoring

* Refactoring

* refactoring

* Added spring boot & hibernate

* Added user & order functional

* added hateoas & pagination for orders

* added audit & refactoring

* audit & dao refactoring

* refactoring & added docs for new classes

* Added tests for new classes & refactoring

* Added more tests

* moved @Transactional from dao implementation to dao tests

* changed @Pre.. annotations to @Post..

* fixed cascade type between Order and GiftCertificate

* fixed sorting and search of most used tag

* changed search predicate from equal to like

* fix after demo

* Added spring security & jwt

* reformat DaoTestData.java

* added docs for new classes & methods

* added tests for new classes & methods

* added exceptions for jwt

* added jwt.properties

* fix comments

* refactor AuthenticationController

+++ model/src/main/java/com/epam/esm/dao/GiftCertificateDao.java
@@ -3,7 +3,6 @@ package com.epam.esm.dao;
 import com.epam.esm.entity.GiftCertificate;
 import com.epam.esm.entity.QueryParameters;
 
-import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -11,11 +10,11 @@ public interface GiftCertificateDao {
 
     List<GiftCertificate> findByParameters(QueryParameters parameters);
 
-    Optional<GiftCertificate> findById(BigInteger id);
+    Optional<GiftCertificate> findById(Long id);
 
     GiftCertificate update(GiftCertificate giftCertificate);
 
-    BigInteger deleteById(BigInteger id);
+    Long deleteById(Long id);
 
     GiftCertificate save(GiftCertificate giftCertificate);
