+++ model/src/main/java/com/epam/esm/dao/impl/GiftCertificateDaoImpl.java
@@ -12,13 +12,16 @@ import com.epam.esm.specification.impl.PaginationSpecificationImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.*;
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.EntityNotFoundException;
+import javax.persistence.PersistenceContext;
+import javax.persistence.PersistenceUnit;
+import javax.persistence.TypedQuery;
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
-import javax.transaction.Transactional;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -67,7 +70,7 @@ public class GiftCertificateDaoImpl implements GiftCertificateDao {
     }
 
     @Override
-    public Optional<GiftCertificate> findById(BigInteger id) {
+    public Optional<GiftCertificate> findById(Long id) {
         GiftCertificate foundGiftCertificate = entityManager.find(GiftCertificate.class, id);
         return Optional.ofNullable(foundGiftCertificate);
     }
@@ -94,7 +97,7 @@ public class GiftCertificateDaoImpl implements GiftCertificateDao {
     }
 
     @Override
-    public BigInteger deleteById(BigInteger id) {
+    public Long deleteById(Long id) {
         try {
             GiftCertificate certificate = entityManager.find(GiftCertificate.class, id);
             if (certificate != null) {
