+++ service/src/main/java/com/epam/esm/service/TagService.java
@@ -5,7 +5,6 @@ import com.epam.esm.dto.TagDto;
 import com.epam.esm.exception.OrderNotFoundException;
 import com.epam.esm.exception.TagNotFoundException;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -17,8 +16,7 @@ public interface TagService {
     /**
      * Retrieves all Order of User with parameters.
      *
-     * @param idUser            id of Order's User. must not be {@literal null}.
-     * @param requestParameters parameters to search orders. must not be {@literal null}.
+     * @param parameters parameters to search orders. must not be {@literal null}.
      * @return list of OrderDto.
      * @throws OrderNotFoundException if Order with given id doesn't exists.
      */
@@ -32,14 +30,14 @@ public interface TagService {
      * @throws TagNotFoundException if Tag with
      *                              given id doesn't exists.
      */
-    TagDto findById(BigInteger id);
+    TagDto findById(Long id);
 
     /**
      * Deletes the Tag with given id.
      *
      * @param id must not be {@literal null}.
      */
-    BigInteger deleteById(BigInteger id);
+    Long deleteById(Long id);
 
     /**
      * Saves a given Tag.
@@ -49,7 +47,7 @@ public interface TagService {
      */
     TagDto save(TagDto tag);
 
-    TagDto getMostWidelyUsedTagOfAUserWithTheHighestCostOfAllOrders(BigInteger idUser);
+    TagDto getMostWidelyUsedTagOfAUserWithTheHighestCostOfAllOrders(Long idUser);
 
     /**
      * Counts number of all Tag records pages.
